+++ src/Middleware/SymfonyController.php
@@ -64,7 +64,10 @@ final class SymfonyController
             $this->symfonyResponse = $this->kernel->handle($this->symfonyRequest);
             $this->symfonyResponse->prepare($this->symfonyRequest);
 
-            return $this->responseConverter->convert($this->symfonyResponse, $connection);
+            $response = $this->responseConverter->convert($this->symfonyResponse, $connection);
+            $response->withProtocolVersion($request->protocolVersion());
+
+            return $response;
         } catch (\Throwable $e) {
             $this->symfonyRequest = null;
             $this->symfonyResponse = null;